requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
redis>=5.0.0
shap>=0.43.0
jinja2>=3.1.0
scikit-learn>=1.3.0
pytest>=7.4.0
httpx>=0.25.0

